Move PreferenceUtils to the commons library
